Reorder imports for domain/staging
As the staging models point to the domain models (specifically, TAS), we need
to re-order the imports. While we're at it, add a `relationship` to the TAS
object.
