+++ dataactcore/models/domainModels.py
@@ -1,7 +1,26 @@
-from dataactcore.models.stagingModels import concatTas
 from sqlalchemy import (
     Column, Date, Index, Integer, Numeric, Text, UniqueConstraint)
 from dataactcore.models.baseModel import Base
+
+
+def concatTas(context):
+    """Create a concatenated TAS string for insert into database."""
+    tas1 = context.current_parameters['allocation_transfer_agency']
+    tas1 = tas1 if tas1 else '000'
+    tas2 = context.current_parameters['agency_identifier']
+    tas2 = tas2 if tas2 else '000'
+    tas3 = context.current_parameters['beginning_period_of_availa']
+    tas3 = tas3 if tas3 else '0000'
+    tas4 = context.current_parameters['ending_period_of_availabil']
+    tas4 = tas4 if tas4 else '0000'
+    tas5 = context.current_parameters['availability_type_code']
+    tas5 = tas5 if tas5 else ' '
+    tas6 = context.current_parameters['main_account_code']
+    tas6 = tas6 if tas6 else '0000'
+    tas7 = context.current_parameters['sub_account_code']
+    tas7 = tas7 if tas7 else '000'
+    tas = '{}{}{}{}{}{}{}'.format(tas1, tas2, tas3, tas4, tas5, tas6, tas7)
+    return tas
 
 
 class TASLookup(Base) :
